调整 AuthenticationBloc 的使用方法
应用启动时，更新 loginUser
登陆成功时，开始推送，且注册 Shortcut
登录失败时，显示提示
登出时，将 loginUser 设置为 null

这个 BLoC 不参与任何页面转跳，与除了 SnackBar 以外的页面显示
